requirements.txt

dj-stripe==2.7.3
Django==4.1.5
djangorestframework==3.14.0
idna==3.4
jsonfield==3.1.0
requests==2.28.1
sslcommerz-lib==1.0
stripe==4.2.0
